fix(MainView): Fix broken busy indicator
Using sprite svg logo icon was not working in Chrome, as in that environment,
CSS selectors which targeted an element in main DOM as well as children
elements in cloned SVG (shadow DOM) were not working.  By using an inline
svg icon widget, this problem goes away, and the busy indicator works in
Chrome, Firefox, and Safari.